Pattern Design Changes
Changed the main modular pattern designed. Suppressed the 'accept' method that was basically useless in the context. Added more automation for the threading context of each module, so that when a module is written, the threads are already managed by the library itself, in order to focus on internal logic rather than the threading model. Also added the possibility to test modules with several config files, and the possibility to use python scripts for data verification in tests